no functional changes, only prefix 2 classes with kate, move one of them from the header to the source, as only used internal
svn path=/trunk/kdelibs/kate/; revision=271426

+++ part/part/katedocument.cpp
@@ -2568,7 +2568,7 @@ bool KateDocument::saveFile()
     {
       int hl (HlManager::self()->detectHighlighting (this));
       
-      if (hl >= 0 && hl != hlMode())
+      if ((hl >= 0) && (uint(hl) != hlMode()))
         internalSetHlMode(hl);
     }
